move reusable_test_data.py from subfolder unit to parent folder tests so the file can be used for selenium tests aswell + change imports of the file accordingly
